concurso_analyzer/requirements.txt
# Dependências principais
crewai>=0.11.0
langchain>=0.1.0
openai
python-dotenv>=1.0.0
pypdf2>=3.0.0
pandas>=2.0.0
matplotlib>=3.0.0
seaborn>=0.12.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
spacy>=3.7.2
nltk>=3.8.1
reportlab

# Dependências de teste
pytest
pytest-cov
coverage

# Ferramentas de desenvolvimento
black
flake8
mypy
